un-nested user class from birdyboard class
