+++ birdyboard.py
@@ -1,5 +1,7 @@
 import uuid
 
+from user import *
+
 
 class Birdyboard:
 
@@ -9,13 +11,6 @@ class Birdyboard:
         self.conversation_list = []
         self.current_user = None
 
-    class User:
-
-        def __init__(self, fullname, screenname):
-            self.screen_name = screenname
-            self.full_name = fullname
-            self.unique_user_ID = str(uuid.uuid1())
-            print(self.unique_user_ID)
 
 
 # if __name__ == '__main__':
